python/requirements.txt
fastapi[all]>=0.75
uvicorn[standard]>=0.15
pytest==8.3.4
pillow==11.1.0
psycopg2>=2
python-multipart>=0.0.20